requirements.txt
aptdaemon==2.0.2
attrs==25.1.0
autocommand==2.2.2
babel==2.17.0
bcc==0.30.0
bcrypt==4.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
Bottleneck==1.4.2
Brlapi==0.8.6
Brotli==1.1.0
certifi==2025.1.31
chardet==5.2.0
click==8.1.8
cloud-init==25.1.1
command-not-found==0.3
configobj==5.0.9
contourpy==1.3.1
coverage==7.6.0
cryptography==43.0.0
cssselect==1.3.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.13
et_xmlfile==2.0.0
fonttools==4.55.3
fs==2.4.16
html5lib-modern==1.2
httplib2==0.22.0
idna==3.10
inflect==7.3.1
iniconfig==1.1.1
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jsonpatch==1.32
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==2.1.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
legacy-cgi==2.6.2
louis==3.32.0
lxml==5.3.2
lz4==4.4.0+dfsg
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
netaddr==1.3.0
netifaces==0.11.0
numexpr==2.10.2
numpy==2.2.3
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.47
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3+dfsg
passlib==1.7.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pycairo==1.27.0
pycups==2.0.4
Pygments==2.18.0
PyGObject==3.50.0
PyJWT==2.10.1
pyparsing==3.1.2
pyserial==3.5
pytest==8.3.5
pytest-cov==5.0.0
python-apt==3.0.0
python-dateutil==2.9.0
python-debian==1.0.1+ubuntu1
python-dotenv==1.0.1
pytz==2025.1
pyxdg==0.28
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
scipy==1.14.1
soupsieve==2.6
sympy==1.13.3
systemd-python==235
tables==3.10.2
tqdm==4.67.1
typeguard==4.4.2
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.17.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.3.0
wadllib==2.0.0
webencodings==0.5.1
wheel==0.45.1
xdg==5
xkit==0.0.0
